+++ src/services/books.api.service.ts
@@ -1,11 +1,11 @@
-import { SearchRequest } from "@/app/books/utils";
+import { SearchRequest } from "@/app/api/books/utils";
 import { BookSearchResult } from "@/components/types";
 
 const searchBooks = async (
   options: SearchRequest
 ): Promise<BookSearchResult> => {
   try {
-    const response = await fetch("/books", {
+    const response = await fetch("/api/books", {
       method: "post",
       body: JSON.stringify(options),
     });
